feat: add ability to upload file in modal window and minor edits
