--- src/api/firebase/updateUserPassword.ts
@@ -1,10 +1,0 @@
-import { updatePassword } from 'firebase/auth';
-
-import { auth } from '@/lib/firebase';
-
-export const updateUserPassword = async (password: string) => {
-  const user = auth.currentUser;
-  if (user && password) {
-    await updatePassword(user, password);
-  }
-};
