+++ src/components/AppRouter/types.ts
@@ -10,7 +10,8 @@ export enum AppRoutes {
   SIGN_UP = '/sign-up',
   LOGIN = '/login',
   PROFILE = '/profile/:id',
-  FEED = '/feed/:id',
+  FEED = '/feed',
+  FEED_TWEET = '/feed/:id',
   EXPLORE = '/explore',
   NOTIFICATIONS = '/notifications',
   MESSAGES = '/messages',
